move (only) all of the files needed by the VisualizerApp
I calculated this list by doing a depth first crawl of the include
statements in both the headers and c++ files. I moved the headers to
either rviz_rendering (defaulting to public headers) or rviz_common
(defaulting to private headers) and changed the extensions to hpp. I
also moved the cpp files to the respective package. I also added a line
in the CMakeLists.txt for all the source files and also the headers if I
determined that they would need Qt's automoc at a glance. I added those
entried commented out so they could be turned on a few at a time when we
start fixing up the code. Aside from changing the extensions on the
header files, the content of the files is untouched.