Switch back to patch instead of git apply.
There were two problems with the fix for the ogre vendor patch:

1.  The downloaded ogre sources is not a git repository
2.  The fix used Unix line endings instead of DOS ones

It seems that newer versions of git do not allow you to call
'git apply' on a directory that is not a git repository.  Worse,
'git apply' silently fails to apply the patch in that case
(actually, in all cases).  So switch back to using patch, which
does properly throw errors and does properly work in a non-git
repository.  Patch fails to apply mixed Unix-style and DOS-style
line endings, so also fix the patch up to use only DOS.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>