Remove uncessary call to render scene
Fixes #486.

At least the call seems unncessary. I haven't noticed any changes in selection render behavior
with or without the change. But, making the call causes an assertion to fail within Ogre when
a selection tool is moused-over an image display (see #486 for details).

The hack was originally added in

https://github.com/ros-visualization/rviz/commit/7feb02d6adf2d3a02343fd43b4b386f515591ab4

but it's not clear if the original buggy behavior still exists without it. There's been a lot
of refactoring since then. I've played around with selecting things and haven't noticed any
flickering or discoloration after removing the hack.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>