Don't install test header files in rviz_rendering. (#564)
* Don't install test header files in rviz_rendering.

This change started as a relatively simple try at not installing
test includes when installing rviz_rendering.  As you can see,
it ballooned into quite a large change, so here is an explanation
of why.

We shouldn't install test include header files when installing
the package; that just ends up on the end user system, and
is a non-supported API.  Worse, we don't want to compile test
code in our main library.  Yet rviz_rendering is currently doing
both of these things.

Unfortunately, it is somewhat tricky to make that code and header
file private.  The problem is that that test code is used in
rviz_rendering, rviz_common, rviz_default_plugins, and
rviz_rendering_tests.

One solution might be to extract that test code into its own
package, and then have all of the other packages depend on it.
The problem is that that test package would both be depended
on by rviz_rendering (for tests), and depends itself on rviz_rendering
(for its functionality), creating a dependency loop.

The solution I opted for here is to copy the test code into the
appropriate packages.  This leads to some duplication of functionality,
but it effectively breaks the dependency loop and succeeds in
eliminating the test code from our installed library.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>