[foxy-backport] fast forward of ros2 (default) to foxy (#570)
* restore compatibility with older Qt versions (#561)

Signed-off-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>

* Suppress warnings when building with older Qt versions. (#562)

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Don't try to moc generate env_config.hpp file. (#550)

This removes one more warning from rviz_common builds.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* rewrite hack to avoid CMake warning with assimp 5.0.1 and older, apply cross platform (#565)

Signed-off-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>

* Use dedicated TransformListener thread (#551)

Signed-off-by: ymd-stella <world.applepie@gmail.com>

* restore alphabetical include order (#563)

Signed-off-by: Karsten Knese <karsten@openrobotics.org>

* Don't install test header files in rviz_rendering. (#564)

* Don't install test header files in rviz_rendering.

This change started as a relatively simple try at not installing
test includes when installing rviz_rendering.  As you can see,
it ballooned into quite a large change, so here is an explanation
of why.

We shouldn't install test include header files when installing
the package; that just ends up on the end user system, and
is a non-supported API.  Worse, we don't want to compile test
code in our main library.  Yet rviz_rendering is currently doing
both of these things.

Unfortunately, it is somewhat tricky to make that code and header
file private.  The problem is that that test code is used in
rviz_rendering, rviz_common, rviz_default_plugins, and
rviz_rendering_tests.

One solution might be to extract that test code into its own
package, and then have all of the other packages depend on it.
The problem is that that test package would both be depended
on by rviz_rendering (for tests), and depends itself on rviz_rendering
(for its functionality), creating a dependency loop.

The solution I opted for here is to copy the test code into the
appropriate packages.  This leads to some duplication of functionality,
but it effectively breaks the dependency loop and succeeds in
eliminating the test code from our installed library.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

Co-authored-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>
Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>
Co-authored-by: ymd-stella <7959916+ymd-stella@users.noreply.github.com>
Co-authored-by: Karsten Knese <Karsten1987@users.noreply.github.com>