Fix a memory leak when using the ResourceIOSystem.
clang static analysis pointed out that when using the "Exists"
method of the ResourceIOSystem, we would leak the newly
created ResourceIOStream.  While I was in there looking at it,
I noticed that the way it went about things was unnecessarily
complicated by the additional structure and 'checkExistsAndOpen'
method.  Instead, just open code the small amount of code
necessary to determine if the file exists (for Exists), and
to actually open the file (for Open).  This simplifies the
code and fixes the memory leak.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>