Silence a dead store warning.
clang static analysis points out that the value stored into
offset is never read near the end of declareVertexBufferOrdering.
While that is technically correct, removing the store is
dangerous; any code that is later added will have to remember
to store it again.  Instead of removing the store, just put
'(void)offset;' at the end to tell clang not to warn about it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>