Expose VisualizationManager and YamlConfigReader to the public API
In order to create a standalone Qt application that contains an embedded RViz2 RenderPanel,
the header for the VisualizationManager class must be moved to the public API to allow
instantiating a VisualizationManager independently from the context of the RViz2 application.
The YamlConfigReader class also needs to be moved to the public API to allow runtime
configuration of the VisualizationManager, for example to add a Grid plugin.

Signed-off-by: Joe Schornak <joe.schornak@gmail.com>