Add ViewPicker::get3DPatch to the public API
ViewPicker::get3DPatch is currently a private member function of the ViewPicker class.
It would be very useful to make it public, since this lets us develop plugins that
can use the orientations of mesh surfaces and point clouds in the 3D scene.
This change moves it from the private API to the public API.

Signed-off-by: Joe Schornak <joe.schornak@gmail.com>