Fix byte indexing for depth patch pixels (#661)
Depth pixel data is now represented by a series of 3-byte sets, rather than a
series of 4-byte sets as it was in ROS1 RViz. This means that the previous method
used to access pixel data by index relative to the start of the depth patch array
does not work correctly, and that each pixel after the first will be composed
incorrectly. This change modifies the pixel indexing method to use sets of 3
bytes instead, and adds an assert to ensure that the depth patch uses the
expected data format before accessing pixel data.

Signed-off-by: Joe Schornak <joe.schornak@gmail.com>