Remove the word "Alpha" from the splash screen. (#696)
I think we can safely say rviz2 is no longer alpha.  This
change adds in the current version from the package.xml, though
this is likely to lag behind when a new release is made.  I'm
not quite sure how to improve that.

While we are in here, also switch from the "ROS 2.0" nomenclature
to just "ROS 2".

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>