Make resource file paths relative
These resource files support relative paths from their directory, so we
should override the default paths for UNIX and WIN32 to act like APPLE,
where relative paths are used by default.

This should be a pretty safe thing to do because the absolute paths
currently in these locations are actually to the staging in the
project's build directory, which isn't available unless building from
source.

Signed-off-by: Scott K Logan <logans@cottsay.net>