Use CMAKE_STAGING_PREFIX for staging OGRE installation (#861)
To maintain build/install phasing correctly, this vendor package is
installing the external project to a staging location, which is then
recursively copied to the final location during the vendor package's
installation phase.

The mechanism by which we're currently telling the external project
where to install to gives it the impression that that will be the final
destination, and it embeds such paths in places like pkgconfig files.

The CMAKE_STAGING_PREFIX can be used to override all of the `install()`
operations' targets while still telling the project where it will
eventually end up, which is pretty much exactly what we're trying to do.

Signed-off-by: Scott K Logan <logans@cottsay.net>
(cherry picked from commit 47e91ee0de52dc4d5b0119120cacbef2595d8e6d)