If vendored assimp is present, always prefer that
The existing behavior is to always prefer the system's assimp package if
it is sufficient, even if the vendored assimp package is present. This
is very confusing behavior when combined with FORCE_BUILD_VENDOR_PKG.

Additionally, the double find_package calls can result in strange
behavior if both calls deal create or modify targets, as targets created
by the first call will be present when the second call tries to create
them.

Signed-off-by: Scott K Logan <logans@cottsay.net>