Remove -DCMAKE_INSTALL_LIBDIR=lib line
As far as we can tell, this is already the default on
Linux.  Building with and without this line in place
on Linux ended up with the same installation directory
structure.

Also, specifying this causes a warning on Windows where
it isn't used.

Since this is just setting the default anyway, remove it.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>