Revamp adding things to the EffortVisual class during setEffort.
This is in an attempt to make sure we don't access
uninitialized memory.  It is a little complicated because
it is possible for the user to have set some of the information
before the setEffort() call, and we don't want to forget that.
So when setEffort() is called, we use the already set values
if they are available.  If they aren't, we use the identity.

On disable, we only erase the effort_arrow and effort_circle,
again so we don't forget information the user may have otherwise
provided us.

With all of this in place, it isn't possible to call setEffort
anymore and access uninitialized memory.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>