Fix ODR violations in interactive_marker displays.
When building this on RHEL and with -Wodr specified,
we get some ODR violations reported.  The main reason
for this is that the InteractiveMarkerNamespaceProperty
class is forward-declared in the wrong namespace.

However, when looking at it I realized that there is
really little reason to forward-declare all of this
stuff at all.  We have header files; we should just
use them.  So I did a minor refactoring here where
I removed forward declarations in favor of the header
file.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>